Finished the example
